CardanoClient

Description

A cardano-node client used in end-to-end tests and benchmarks.

This modules contains some more functions besides the re-exported basic querying of hydra-node's CardanoClient.

Documentation

module Hydra.Chain.CardanoClient

data CardanoClientException

Constructors

ClientQueryException QueryException

ClientSubmitException

Fields

reason :: Text

tx :: Tx

data Sizes

Constructors

Sizes

Fields

inputs :: Int

outputs :: Int

witnesses :: Int

buildAddress :: VerificationKey PaymentKey -> NetworkId -> Address ShelleyAddr

Build an address give a key.

From https://github.com/input-output-hk/cardano-node/blob/master/cardano-cli/src/Cardano/CLI/Shelley/Run/Address.hs#L106 Throws CardanoClientException if the query fails.

buildScriptAddress :: Script -> NetworkId -> Address ShelleyAddr

queryStakePools :: NetworkId -> FilePath -> QueryPoint -> IO (Set PoolId)

buildRaw :: [TxIn] -> [TxOut CtxTx] -> Lovelace -> Either TxBodyError TxBody

Build a "raw" transaction from a bunch of inputs, outputs and fees.

build :: NetworkId -> FilePath -> Address ShelleyAddr -> [(TxIn, Maybe (Script, ScriptData, ScriptRedeemer))] -> [TxIn] -> [TxOut CtxTx] -> IO (Either TxBodyErrorAutoBalance TxBody)

calculateMinFee :: NetworkId -> TxBody -> Sizes -> ProtocolParameters -> Lovelace

defaultSizes :: Sizes

sign :: SigningKey PaymentKey -> TxBody -> Tx

Sign a transaction body with given signing key.

submit :: NetworkId -> FilePath -> Tx -> IO ()

Submit a (signed) transaction to the node.

Throws CardanoClientException if submission fails.

waitForPayment :: NetworkId -> FilePath -> Lovelace -> Address ShelleyAddr -> IO UTxO

waitForUTxO :: NetworkId -> FilePath -> UTxO -> IO ()

waitForTransaction :: NetworkId -> FilePath -> Tx -> IO UTxO

mkGenesisTx

Arguments

:: NetworkId

-> ProtocolParameters

-> SigningKey PaymentKey
Owner of the initialFund.

-> Lovelace
Amount of initialFunds

-> [(VerificationKey PaymentKey, Lovelace)]
Recipients and amounts to pay in this transaction.

-> Tx