CardanoCluster

Documentation

defaultNetworkId :: NetworkId

TODO: This is hard-coded and must match what's in the genesis file, so ideally, we want to either:

overwrite the genesis configuration with the ClusterConfig

pull the network id from the genesis configuration

availableInitialFunds :: Num a => a

data Actor

Enumeration of known actors for which we can get the keysFor and writeKeysFor.

Constructors

Alice

Bob

Carol

Faucet

actorName :: Actor -> String

keysFor :: Actor -> IO (VerificationKey PaymentKey, SigningKey PaymentKey)

Get the "well-known" keys for given actor.

fromRawVKey :: VerificationKey -> VerificationKey PaymentKey

chainConfigFor :: Actor -> FilePath -> FilePath -> [Actor] -> IO ChainConfig

Starting a cluster or single nodes

data RunningCluster

Constructors

RunningCluster ClusterConfig [RunningNode]

data ClusterConfig

Configuration parameters for the cluster.

Constructors

ClusterConfig

Fields

parentStateDirectory :: FilePath

networkId :: NetworkId

asSigningKey :: AsType (SigningKey PaymentKey)

withCluster :: Tracer IO ClusterLog -> ClusterConfig -> (RunningCluster -> IO ()) -> IO ()

data Marked

Constructors

Fuel

Normal

seedFromFaucet

Arguments

:: NetworkId

-> RunningNode

-> VerificationKey PaymentKey
Recipient of the funds

-> Lovelace
Amount to get from faucet

-> Marked
Marked as fuel or normal output?

-> IO UTxO

Create a specially marked "seed" UTXO containing requested Lovelace by redeeming funds available to the well-known faucet.

NOTE: This function is querying and looping forever until it finds a suitable output!

seedFromFaucet_

Arguments

:: NetworkId

-> RunningNode

-> VerificationKey PaymentKey
Recipient of the funds

-> Lovelace
Amount to get from faucet

-> Marked
Marked as fuel or normal output?

-> IO ()

Like seedFromFaucet, but without returning the seeded UTxO.

data ClusterLog

Constructors

MsgFromNode NodeId NodeLog

MsgNodeStarting CardanoNodeConfig