CardanoNode

Documentation

type Port = Int

newtype NodeId

Constructors

NodeId Int

data RunningNode

Constructors

RunningNode NodeId FilePath

data CardanoNodeConfig

Configuration parameters for a single node of the cluster

Constructors

CardanoNodeConfig

Fields

nodeId :: NodeId

An identifier for the node

stateDirectory :: FilePath

Parent state directory in which create a state directory for the cluster

systemStart :: UTCTime

Blockchain start time

ports :: PortsConfig

A list of port

newNodeConfig :: FilePath -> IO CardanoNodeConfig

data CardanoNodeArgs

Arguments given to the 'cardano-node' command-line to run a node.

Constructors

CardanoNodeArgs

Fields

nodeSocket :: FilePath

nodeConfigFile :: FilePath

nodeByronGenesisFile :: FilePath

nodeShelleyGenesisFile :: FilePath

nodeAlonzoGenesisFile :: FilePath

nodeTopologyFile :: FilePath

nodeDatabaseDir :: FilePath

nodeDlgCertFile :: Maybe FilePath

nodeSignKeyFile :: Maybe FilePath

nodeOpCertFile :: Maybe FilePath

nodeKesKeyFile :: Maybe FilePath

nodeVrfKeyFile :: Maybe FilePath

nodePort :: Maybe Port

defaultCardanoNodeArgs :: CardanoNodeArgs

data PortsConfig

Configuration of ports from the perspective of a peer in the context of a fully sockected topology.

Constructors

PortsConfig

Fields

ours :: Port

Our node TCP port.

peers :: [Port]

Other peers TCP ports.

getCardanoNodeVersion :: IO String

withCardanoNodeDevnet :: Tracer IO NodeLog -> CardanoNodeConfig -> (RunningNode -> IO ()) -> IO ()

Start a single cardano-node devnet using the config from config/ and credentials from configcredentials. Only the Faucet actor will receive "initialFunds". Use seedFromFaucet to distribute funds other wallets.

withCardanoNodeOnKnownNetwork :: Tracer IO NodeLog -> FilePath -> KnownNetwork -> (RunningNode -> IO ()) -> IO ()

Run a cardano-node as normal network participant on a known network.

waitForFullySynchronized :: Tracer IO NodeLog -> KnownNetwork -> RunningNode -> IO ()

Wait until the node is fully caught up with the network. This can take a while!

withCardanoNode :: Tracer IO NodeLog -> CardanoNodeConfig -> CardanoNodeArgs -> (RunningNode -> IO ()) -> IO ()

waitForSocket :: RunningNode -> IO ()

Wait for the node socket file to become available.

cardanoNodeProcess :: Maybe FilePath -> CardanoNodeArgs -> CreateProcess

Generate command-line arguments for launching cardano-node.

initSystemStart :: IO UTCTime

Initialize the system start time to now (modulo a small offset needed to give time to the system to bootstrap correctly).

refreshSystemStart :: CardanoNodeConfig -> CardanoNodeArgs -> IO ()

Re-generate configuration and genesis files with fresh system start times.

mkTopology :: [Port] -> Value

Generate a topology file from a list of peers.

generateCardanoKey :: IO (VerificationKey PaymentKey, SigningKey PaymentKey)

data ProcessHasExited

Constructors

ProcessHasExited Text ExitCode

data NodeLog

Constructors

MsgUsingStateDirectory FilePath

MsgNodeCmdSpec Text

MsgCLI [Text]

MsgCLIStatus Text Text

MsgCLIRetry Text

MsgCLIRetryResult Text Int

MsgNodeStarting CardanoNodeConfig

MsgSocketIsReady FilePath

MsgSynchronizing

Fields

percentDone :: Centi

MsgNodeIsReady

addField :: ToJSON a => Key -> a -> Value -> Value

withObject :: (Object -> Object) -> Value -> Value

Do something with an a JSON object. Fails if the given JSON value isn't an object.

unsafeDecodeJsonFile :: FromJSON a => FilePath -> IO a