Hydra.Cluster.Faucet

Documentation

data Marked

Constructors

Fuel

Normal

seedFromFaucet

Arguments

:: NetworkId

-> RunningNode

-> VerificationKey PaymentKey
Recipient of the funds

-> Lovelace
Amount to get from faucet

-> Marked
Marked as fuel or normal output?

-> IO UTxO

Create a specially marked "seed" UTXO containing requested Lovelace by redeeming funds available to the well-known faucet.

NOTE: This function is querying and looping forever until it finds a suitable output!

seedFromFaucet_

Arguments

:: NetworkId

-> RunningNode

-> VerificationKey PaymentKey
Recipient of the funds

-> Lovelace
Amount to get from faucet

-> Marked
Marked as fuel or normal output?

-> IO ()

Like seedFromFaucet, but without returning the seeded UTxO.