Hydra.Cluster.Fixture

Description

Example values and utilities used across hydra-cluster tests and benchmarks.

Documentation

alice :: Party

bob :: Party

carol :: Party

aliceSk :: SigningKey

bobSk :: SigningKey

carolSk :: SigningKey

aliceVk :: VerificationKey

bobVk :: VerificationKey

carolVk :: VerificationKey

cperiod :: ContestationPeriod

defaultNetworkId :: NetworkId

TODO: This is hard-coded and must match what's in the genesis file, so ideally, we want to either:

overwrite the genesis configuration with the ClusterConfig

pull the network id from the genesis configuration

availableInitialFunds :: Num a => a

data Actor

Enumeration of known actors for which we can get the keysFor and writeKeysFor.

Constructors

Alice

Bob

Carol

Faucet

actorName :: Actor -> String

data KnownNetwork

A network known to the hydra-cluster. That means we have configuration files to connect to at least these networks.

Constructors

Testnet

VasilTestnet

knownNetworkId :: KnownNetwork -> NetworkId