Hydra.Cluster.Util

Description

Utilities used across hydra-cluster

Documentation

readConfigFile :: FilePath -> IO ByteString

Lookup a config file similar reading a file from disk. If the env variable HYDRA_CONFIG_DIR is set, filenames will be resolved relative to its value otherwise they will be looked up in the package's data path.

keysFor :: Actor -> IO (VerificationKey PaymentKey, SigningKey PaymentKey)

Get the "well-known" keys for given actor.

createAndSaveSigningKey :: FilePath -> IO (SigningKey PaymentKey)

Create and save new signing key at the provided path. NOTE: Uses TextEnvelope format.

chainConfigFor :: HasCallStack => Actor -> FilePath -> FilePath -> [Actor] -> ContestationPeriod -> IO ChainConfig