Hydra.Generator

Documentation

networkId :: NetworkId

data Dataset

A Dataset that can be run for testing purpose. Each Dataset represents a complete scenario where several ClientDataset are run concurrently against one or more HydraNodes. A dataset can optionally have a title and description which will be used to report results.

Constructors

Dataset

Fields

fundingTransaction :: Tx

clientDatasets :: [ClientDataset]

title :: Maybe Text

description :: Maybe Text

data ClientKeys

Constructors

ClientKeys

Fields

signingKey :: SigningKey PaymentKey

Key used by the hydra-node to authorize hydra transactions and holding fuel.

externalSigningKey :: SigningKey PaymentKey

Key holding funds to commit.

data ClientDataset

Constructors

ClientDataset

Fields

clientKeys :: ClientKeys

initialUTxO :: UTxO

txSequence :: [Tx]

defaultProtocolParameters :: ProtocolParameters

generateConstantUTxODataset

Arguments

:: ProtocolParameters

-> Int
Number of clients

-> Int
Number of transactions

-> IO Dataset

Generate Dataset which does not grow the per-client UTXO set over time. The sequence of transactions generated consist only of simple payments from and to arbitrary keys controlled by the individual clients.

genDatasetConstantUTxO

Arguments

:: SigningKey PaymentKey
The faucet signing key

-> ProtocolParameters

-> Int
Number of clients

-> Int
Number of transactions

-> Gen Dataset