HydraNode

Documentation

data HydraClient

Constructors

HydraClient

Fields

hydraNodeId :: Int

connection :: Connection

tracer :: Tracer IO EndToEndLog

withHydraNode :: Tracer IO EndToEndLog -> ChainConfig -> FilePath -> Int -> SigningKey -> [VerificationKey] -> [Int] -> (HydraClient -> IO a) -> IO a

send :: HydraClient -> Value -> IO ()

input :: Text -> [Pair] -> Value

Create an input as expected by send.

waitFor :: HasCallStack => Tracer IO EndToEndLog -> Natural -> [HydraClient] -> Value -> IO ()

Wait some time for a single output from each of given nodes. This function waits for delay seconds for message expected to be seen by all given nodes.

waitMatch :: HasCallStack => Natural -> HydraClient -> (Value -> Maybe a) -> IO a

output :: Text -> [Pair] -> Value

Create an output as expected by waitFor and waitForAll.

getMetrics :: HasCallStack => HydraClient -> IO ByteString

queryNode :: Int -> IO (Response ByteString)

hydraNodeProcess :: Options -> CreateProcess

waitForProcess :: ProcessHandle -> IO ExitCode

Waits for the specified process to terminate, and returns its exit code.

GHC Note: in order to call waitForProcess without blocking all the other threads in the system, you must compile the program with -threaded.

Note that it is safe to call waitForProcess for the same process in multiple threads. When the process ends, threads blocking on this call will wake in FIFO order.

(Since: 1.2.0.0) On Unix systems, a negative value ExitFailure -signum indicates that the child was terminated by signal signum. The signal numbers are platform-specific, so to test for a specific signal use the constants provided by System.Posix.Signals in the unix package. Note: core dumps are not reported, use System.Posix.Process if you need this detail.

readCreateProcess

Arguments

:: CreateProcess

-> String
standard input

-> IO String
stdout

readCreateProcess works exactly like readProcess except that it lets you pass CreateProcess giving better flexibility.

 > readCreateProcess ((shell "pwd") { cwd = Just "/etc/" }) ""
 "/etc\n"

Note that Handles provided for std_in or std_out via the CreateProcess record will be ignored.

Since: process-1.2.3.0

withCreateProcess :: CreateProcess -> (Maybe Handle -> Maybe Handle -> Maybe Handle -> ProcessHandle -> IO a) -> IO a

A bracket-style resource handler for createProcess.

Does automatic cleanup when the action finishes. If there is an exception in the body then it ensures that the process gets terminated and any CreatePipe Handles are closed. In particular this means that if the Haskell thread is killed (e.g. killThread), that the external process is also terminated.

e.g.

withCreateProcess (proc cmd args) { ... }  $ \stdin stdout stderr ph -> do
  ...

Since: process-1.4.3.0

proc :: FilePath -> [String] -> CreateProcess

Construct a CreateProcess record for passing to createProcess, representing a raw command with arguments.

See RawCommand for precise semantics of the specified FilePath.

data CreateProcess

Constructors

CreateProcess

Fields

cmdspec :: CmdSpec

Executable & arguments, or shell command. If cwd is Nothing, relative paths are resolved with respect to the current working directory. If cwd is provided, it is implementation-dependent whether relative paths are resolved with respect to cwd or the current working directory, so absolute paths should be used to ensure portability.

cwd :: Maybe FilePath

Optional path to the working directory for the new process

env :: Maybe [(String, String)]

Optional environment (otherwise inherit from the current process)

std_in :: StdStream

How to determine stdin

std_out :: StdStream

How to determine stdout

std_err :: StdStream

How to determine stderr

close_fds :: Bool

Close all file descriptors except stdin, stdout and stderr in the new process (on Windows, only works if std_in, std_out, and std_err are all Inherit). This implementation will call close an every fd from 3 to the maximum of open files, which can be slow for high maximum of open files.

create_group :: Bool

Create a new process group

delegate_ctlc :: Bool

Delegate control-C handling. Use this for interactive console processes to let them handle control-C themselves (see below for details).

On Windows this has no effect.

Since: process-1.2.0.0

detach_console :: Bool

Use the windows DETACHED_PROCESS flag when creating the process; does nothing on other platforms.

Since: process-1.3.0.0

create_new_console :: Bool

Use the windows CREATE_NEW_CONSOLE flag when creating the process; does nothing on other platforms.

Default: False

Since: process-1.3.0.0

new_session :: Bool

Use posix setsid to start the new process in a new session; does nothing on other platforms.

Since: process-1.3.0.0

child_group :: Maybe GroupID

Use posix setgid to set child process's group id; does nothing on other platforms.

Default: Nothing

Since: process-1.4.0.0

child_user :: Maybe UserID

Use posix setuid to set child process's user id; does nothing on other platforms.

Default: Nothing

Since: process-1.4.0.0

use_process_jobs :: Bool

On Windows systems this flag indicates that we should wait for the entire process tree to finish before unblocking. On POSIX systems this flag is ignored. See $exec-on-windows for details.

Default: False

Since: process-1.5.0.0

data StdStream

Constructors

Inherit
Inherit Handle from parent

UseHandle Handle
Use the supplied Handle

CreatePipe
Create a new pipe. The returned Handle will use the default encoding and newline translation mode (just like Handles created by openFile).

NoStream
Close the stream's file descriptor without passing a Handle. On POSIX systems this may lead to strange behavior in the child process because attempting to read or write after the file has been closed throws an error. This should only be used with child processes that don't use the file descriptor at all. If you wish to ignore the child process's output you should either create a pipe and drain it manually or pass a Handle that writes to /dev/null.

data ProcessHandle

A handle to a process, which can be used to wait for termination of the process using waitForProcess.

None of the process-creation functions in this library wait for termination: they all return a ProcessHandle which may be used to wait for the process later.

On Windows a second wait method can be used to block for event completion. This requires two handles. A process job handle and a events handle to monitor.

waitForNodesConnected :: HasCallStack => Tracer IO EndToEndLog -> [HydraClient] -> IO ()

waitNext :: HasCallStack => HydraClient -> IO Value

withNewClient :: HydraClient -> (HydraClient -> IO a) -> IO a

Runs an action with a new connection to given Hydra node.

withHydraCluster

Arguments

:: Tracer IO EndToEndLog

-> FilePath

-> FilePath

-> Int
First node id

-> [(VerificationKey PaymentKey, SigningKey PaymentKey)]
NOTE: This decides on the size of the cluster!

-> [SigningKey]

-> (NonEmpty HydraClient -> IO ())

-> IO ()

data EndToEndLog

Constructors

NodeStarted Int

SentMessage Int Value

StartWaiting [Int] [Value]

ReceivedMessage Int Value

EndWaiting Int

FromCardanoNode NodeLog

StartingFunds

Fields

actor :: String

fuelUTxO :: UTxO

otherUTxO :: UTxO

RefueledFunds

Fields

actor :: String

refuelingAmount :: Lovelace

fuelUTxO :: UTxO

RemainingFunds

Fields

actor :: String

fuelUTxO :: UTxO

otherUTxO :: UTxO