Hydra.Cluster.Faucet

Documentation

data Marked

Constructors

Fuel

Normal

data FaucetException

Constructors

FaucetHasNotEnoughFunds

Fields

faucetUTxO :: UTxO

FaucetFailedToBuildTx

Fields

reason :: TxBodyErrorAutoBalance

data FaucetLog

Constructors

TraceResourceExhaustedHandled Text

ReturnedFunds

Fields

actor :: String

returnAmount :: Lovelace

seedFromFaucet

Arguments

:: RunningNode

-> VerificationKey PaymentKey
Recipient of the funds

-> Lovelace
Amount to get from faucet

-> Marked
Marked as fuel or normal output?

-> Tracer IO FaucetLog

-> IO UTxO

Create a specially marked "seed" UTXO containing requested Lovelace by redeeming funds available to the well-known faucet.

findFaucetUTxO :: RunningNode -> Lovelace -> IO UTxO

seedFromFaucet_

Arguments

:: RunningNode

-> VerificationKey PaymentKey
Recipient of the funds

-> Lovelace
Amount to get from faucet

-> Marked
Marked as fuel or normal output?

-> Tracer IO FaucetLog

-> IO ()

Like seedFromFaucet, but without returning the seeded UTxO'.

returnFundsToFaucet :: Tracer IO FaucetLog -> RunningNode -> Actor -> IO ()

Return the remaining funds to the faucet

createOutputAtAddress :: ToScriptData a => RunningNode -> ProtocolParameters -> AddressInEra -> a -> IO (TxIn, TxOut CtxUTxO)

calculateTxFee :: RunningNode -> SigningKey PaymentKey -> UTxO -> AddressInEra -> Lovelace -> IO Lovelace

Build and sign tx and return the calculated fee. - Signing key should be the key of a sender - Address is used as a change address.

retryOnExceptions :: (MonadCatch m, MonadDelay m) => Tracer m FaucetLog -> m () -> m ()

Try to submit tx and retry when some caught exception/s take place.

publishHydraScriptsAs :: RunningNode -> Actor -> IO TxId

Publish current Hydra scripts as scripts outputs for later referencing them.

The key of the given Actor is used to pay for fees in required transactions, it is expected to have sufficient funds.

queryMarkedUTxO :: RunningNode -> VerificationKey PaymentKey -> IO (UTxO, UTxO)

Like queryUTxOFor at the tip, but also partition outputs marked as Fuel and Normal.

Throws at least QueryException if query fails.