Hydra.Cluster.Faucet

Documentation

data Marked

Constructors

Fuel

Normal

data FaucetException

Constructors

FaucetHasNotEnoughFunds

Fields

faucetUTxO :: UTxO

FaucetFailedToBuildTx

Fields

reason :: TxBodyErrorAutoBalance

newtype FaucetLog

Constructors

TraceResourceExhaustedHandled Text

seedFromFaucet

Arguments

:: RunningNode

-> VerificationKey PaymentKey
Recipient of the funds

-> Lovelace
Amount to get from faucet

-> Marked
Marked as fuel or normal output?

-> Tracer IO FaucetLog

-> IO UTxO

Create a specially marked "seed" UTXO containing requested Lovelace by redeeming funds available to the well-known faucet.

seedFromFaucet_

Arguments

:: RunningNode

-> VerificationKey PaymentKey
Recipient of the funds

-> Lovelace
Amount to get from faucet

-> Marked
Marked as fuel or normal output?

-> Tracer IO FaucetLog

-> IO ()

Like seedFromFaucet, but without returning the seeded UTxO.

publishHydraScriptsAs :: RunningNode -> Actor -> IO TxId

Publish current Hydra scripts as scripts outputs for later referencing them.

The key of the given Actor is used to pay for fees in required transactions, it is expected to have sufficient funds.

queryMarkedUTxO :: RunningNode -> VerificationKey PaymentKey -> IO (UTxO, UTxO)

Like queryUTxOFor at the tip, but also partition outputs marked as Fuel and Normal.

Throws at least QueryException if query fails.