Hydra.Cluster.Faucet

Documentation

data Marked

Constructors

Fuel

Normal

data FaucetException

Constructors

FaucetHasNotEnoughFunds

Fields

faucetUTxO :: UTxO

FaucetFailedToBuildTx

Fields

reason :: TxBodyErrorAutoBalance

seedFromFaucet

Arguments

:: NetworkId

-> RunningNode

-> VerificationKey PaymentKey
Recipient of the funds

-> Lovelace
Amount to get from faucet

-> Marked
Marked as fuel or normal output?

-> IO UTxO

Create a specially marked "seed" UTXO containing requested Lovelace by redeeming funds available to the well-known faucet.

seedFromFaucet_

Arguments

:: NetworkId

-> RunningNode

-> VerificationKey PaymentKey
Recipient of the funds

-> Lovelace
Amount to get from faucet

-> Marked
Marked as fuel or normal output?

-> IO ()

Like seedFromFaucet, but without returning the seeded UTxO.

queryUTxOFor :: NetworkId -> RunningNode -> QueryPoint -> VerificationKey PaymentKey -> IO UTxO

Query UTxO for the address of given verification key at point.

Throws at least QueryException if query fails.

queryMarkedUTxO :: NetworkId -> RunningNode -> VerificationKey PaymentKey -> IO (UTxO, UTxO)

Like queryUTxOFor at the tip, but also partition outputs marked as Fuel and Normal.

Throws at least QueryException if query fails.