Hydra.Cluster.Scenarios

Documentation

restartedNodeCanObserveCommitTx :: Tracer IO EndToEndLog -> FilePath -> RunningNode -> TxId -> IO ()

restartedNodeCanAbort :: Tracer IO EndToEndLog -> FilePath -> RunningNode -> TxId -> IO ()

singlePartyHeadFullLifeCycle :: Tracer IO EndToEndLog -> FilePath -> RunningNode -> TxId -> IO ()

Step through the full life cycle of a Hydra Head with only a single participant. This scenario is also used by the smoke test run via the `hydra-cluster` executable.

singlePartyCommitsUsingFuel :: Tracer IO EndToEndLog -> FilePath -> RunningNode -> TxId -> IO ()

Ensures the _old_ way of committing (using Fuel) still works.

singlePartyCommitsFromExternalScript :: Tracer IO EndToEndLog -> FilePath -> RunningNode -> TxId -> IO ()

Single hydra-node where the commit is done from an external UTxO owned by a script which requires providing script, datum and redeemer instead of signing the transaction.

singlePartyCannotCommitExternallyWalletUtxo :: Tracer IO EndToEndLog -> FilePath -> RunningNode -> TxId -> IO ()

canCloseWithLongContestationPeriod :: Tracer IO EndToEndLog -> FilePath -> RunningNode -> TxId -> IO ()

Initialize open and close a head on a real network and ensure contestation period longer than the time horizon is possible. For this it is enough that we can close a head and not wait for the deadline.

refuelIfNeeded :: Tracer IO EndToEndLog -> RunningNode -> Actor -> Lovelace -> IO ()

Refuel given Actor with given Lovelace if current marked UTxO is below that amount.

returnFundsToFaucet :: Tracer IO EndToEndLog -> RunningNode -> Actor -> IO ()

Return the remaining funds to the faucet

headIsInitializingWith :: Set Party -> Value -> Maybe HeadId