Hydra.Generator

Documentation

networkId :: NetworkId

data Dataset

A Dataset that can be run for testing purpose. The transactionSequence is guaranteed to be applicable, in sequence, to the initialUTxO set.

Constructors

Dataset

Fields

fundingTransaction :: Tx

clientDatasets :: [ClientDataset]

data ClientDataset

Constructors

ClientDataset

Fields

signingKey :: SigningKey PaymentKey

initialUTxO :: UTxO

txSequence :: [Tx]

defaultProtocolParameters :: ProtocolParameters

generateConstantUTxODataset

Arguments

:: ProtocolParameters

-> Int
Number of clients

-> Int
Number of transactions

-> IO Dataset

Generate Dataset which does not grow the per-client UTXO set over time. The sequence of transactions generated consist only of simple payments from and to arbitrary keys controlled by the individual clients.

genDatasetConstantUTxO

Arguments

:: SigningKey PaymentKey
The faucet signing key

-> ProtocolParameters

-> Int
Number of clients

-> Int
Number of transactions

-> Gen Dataset