HydraNode

Documentation

data HydraClient

Constructors

HydraClient

Fields

hydraNodeId :: Int

connection :: Connection

tracer :: Tracer IO EndToEndLog

input :: Text -> [Pair] -> Value

Create an input as expected by send.

send :: HydraClient -> Value -> IO ()

waitNext :: HasCallStack => HydraClient -> IO Value

output :: Text -> [Pair] -> Value

Create an output as expected by waitFor and waitForAll.

waitFor :: HasCallStack => Tracer IO EndToEndLog -> NominalDiffTime -> [HydraClient] -> Value -> IO ()

Wait some time for a single API server output from each of given nodes. This function waits for delay seconds for message expected to be seen by all given nodes.

waitMatch :: HasCallStack => NominalDiffTime -> HydraClient -> (Value -> Maybe a) -> IO a

Wait up to some time for an API server output to match the given predicate.

waitForAllMatch :: (Eq a, Show a, HasCallStack) => NominalDiffTime -> [HydraClient] -> (Value -> Maybe a) -> IO a

Wait up to some delay for some JSON Value to match given function.

This is a generalisation of waitMatch to multiple nodes.

waitForAll :: HasCallStack => Tracer IO EndToEndLog -> NominalDiffTime -> [HydraClient] -> [Value] -> IO ()

Wait some time for a list of outputs from each of given nodes. This function is the generalised version of waitFor, allowing several messages to be waited for and received in any order.

externalCommit' :: HydraClient -> UTxO' TxOutWithWitness -> IO Tx

Create a commit tx using the hydra-node for later submission

externalCommit :: HydraClient -> UTxO -> IO Tx

Helper to make it easy to externally commit non-script utxo

getMetrics :: HasCallStack => HydraClient -> IO ByteString

data EndToEndLog

Constructors

NodeStarted

Fields

nodeId :: Int

SentMessage

Fields

nodeId :: Int

message :: Value

StartWaiting

Fields

nodeIds :: [Int]

messages :: [Value]

ReceivedMessage

Fields

nodeId :: Int

message :: Value

EndWaiting

Fields

nodeId :: Int

FromCardanoNode NodeLog

FromFaucet FaucetLog

StartingFunds

Fields

actor :: String

fuelUTxO :: UTxO

otherUTxO :: UTxO

RefueledFunds

Fields

actor :: String

refuelingAmount :: Lovelace

fuelUTxO :: UTxO

RemainingFunds

Fields

actor :: String

fuelUTxO :: UTxO

otherUTxO :: UTxO

PublishedHydraScriptsAt

Fields

hydraScriptsTxId :: TxId

UsingHydraScriptsAt

Fields

hydraScriptsTxId :: TxId

withHydraCluster

Arguments

:: HasCallStack

=> Tracer IO EndToEndLog

-> FilePath

-> FilePath

-> Int
First node id This sets the starting point for assigning ports

-> [(VerificationKey PaymentKey, SigningKey PaymentKey)]
NOTE: This decides on the size of the cluster!

-> [SigningKey HydraKey]

-> TxId
Transaction id at which Hydra scripts should have been published.

-> ContestationPeriod

-> (NonEmpty HydraClient -> IO a)

-> IO a

withHydraNode

Arguments

:: Tracer IO EndToEndLog

-> ChainConfig

-> FilePath

-> Int

-> SigningKey HydraKey

-> [VerificationKey HydraKey]

-> [Int]

-> TxId
Transaction id at which Hydra scripts should have been published.

-> (HydraClient -> IO a)

-> IO a

Run a hydra-node with given ChainConfig and using the config from config/.

withConnectionToNode :: Tracer IO EndToEndLog -> Int -> (HydraClient -> IO a) -> IO a

withNewClient :: HydraClient -> (HydraClient -> IO a) -> IO a

Runs an action with a new connection to given Hydra node.

hydraNodeProcess :: RunOptions -> CreateProcess

waitForNodesConnected :: HasCallStack => Tracer IO EndToEndLog -> [HydraClient] -> IO ()