HydraNode

Documentation

data HydraClient

Constructors

HydraClient

Fields

hydraNodeId :: Int

connection :: Connection

tracer :: Tracer IO EndToEndLog

withHydraNode

Arguments

:: Tracer IO EndToEndLog

-> ChainConfig

-> FilePath

-> Int

-> SigningKey HydraKey

-> [VerificationKey HydraKey]

-> [Int]

-> TxId
Transaction id at which Hydra scripts should have been published.

-> (HydraClient -> IO a)

-> IO a

Run a hydra-node with given ChainConfig and using the config from config. Note that this is using the 'configdevnet/genesis-shelly.json' irrelevant of the used cardano network.

send :: HydraClient -> Value -> IO ()

input :: Text -> [Pair] -> Value

Create an input as expected by send.

waitFor :: HasCallStack => Tracer IO EndToEndLog -> Natural -> [HydraClient] -> Value -> IO ()

Wait some time for a single API server output from each of given nodes. This function waits for delay seconds for message expected to be seen by all given nodes.

waitMatch :: HasCallStack => Natural -> HydraClient -> (Value -> Maybe a) -> IO a

output :: Text -> [Pair] -> Value

Create an output as expected by waitFor and waitForAll.

getMetrics :: HasCallStack => HydraClient -> IO ByteString

queryNode :: Int -> IO (Response ByteString)

hydraNodeProcess :: RunOptions -> CreateProcess

data ProcessHandle

proc :: FilePath -> [String] -> CreateProcess

readCreateProcess :: CreateProcess -> String -> IO String

waitForProcess :: ProcessHandle -> IO ExitCode

withCreateProcess :: CreateProcess -> (Maybe Handle -> Maybe Handle -> Maybe Handle -> ProcessHandle -> IO a) -> IO a

data CreateProcess

Constructors

CreateProcess

Fields

cmdspec :: CmdSpec

cwd :: Maybe FilePath

env :: Maybe [(String, String)]

std_in :: StdStream

std_out :: StdStream

std_err :: StdStream

close_fds :: Bool

create_group :: Bool

delegate_ctlc :: Bool

detach_console :: Bool

create_new_console :: Bool

new_session :: Bool

child_group :: Maybe GroupID

child_user :: Maybe UserID

use_process_jobs :: Bool

data StdStream

Constructors

Inherit

UseHandle Handle

CreatePipe

NoStream

waitForNodesConnected :: HasCallStack => Tracer IO EndToEndLog -> [HydraClient] -> IO ()

waitNext :: HasCallStack => HydraClient -> IO Value

withNewClient :: HydraClient -> (HydraClient -> IO a) -> IO a

Runs an action with a new connection to given Hydra node.

withHydraCluster

Arguments

:: HasCallStack

=> Tracer IO EndToEndLog

-> FilePath

-> FilePath

-> Int
First node id

-> [(VerificationKey PaymentKey, SigningKey PaymentKey)]
NOTE: This decides on the size of the cluster!

-> [SigningKey HydraKey]

-> TxId
Transaction id at which Hydra scripts should have been published.

-> (NonEmpty HydraClient -> IO ())

-> IO ()

data EndToEndLog

Constructors

NodeStarted

Fields

nodeId :: Int

SentMessage

Fields

nodeId :: Int

message :: Value

StartWaiting

Fields

nodeIds :: [Int]

messages :: [Value]

ReceivedMessage

Fields

nodeId :: Int

message :: Value

EndWaiting

Fields

nodeId :: Int

FromCardanoNode NodeLog

FromFaucet FaucetLog

StartingFunds

Fields

actor :: String

fuelUTxO :: UTxO

otherUTxO :: UTxO

RefueledFunds

Fields

actor :: String

refuelingAmount :: Lovelace

fuelUTxO :: UTxO

RemainingFunds

Fields

actor :: String

fuelUTxO :: UTxO

otherUTxO :: UTxO

PublishedHydraScriptsAt

Fields

hydraScriptsTxId :: TxId

UsingHydraScriptsAt

Fields

hydraScriptsTxId :: TxId